requirements.txt
gspread==3.6.0
discord.py==1.4.1
